Fix build failure revealed by 75279d7ae04c449c51c9d10c1d4d7f7cdf61c52b
This commit resolves a Linux kernel build failure that was revealed by
75279d7ae04c449c51c9d10c1d4d7f7cdf61c52b. The patch introduces two new
intrinsics, which ultimately changes the intrinsic numbering of other PPC
intrinsics. This causes an issue introduced by
c419cfde48190eefd035420696e28dea4a8371ce, as the patch checks for intrinsics
with particular values, but the addition of the fnabs/fnabss intrinsics updates
the original sqrt/sdiv intrinsic values.