Revert "Round up zero-sized symbols to 1 byte in `.debug_aranges`."
This reverts commit 14c759badafac887246cd62fafdb7f854b848908 (and
also the follow-up commit 151636c7aceccda07976268 that moved a test
case to a different directory).

As discussed in https://reviews.llvm.org/D126257 there is a suspicion
that something was wrong with this commit as text section range was
shortened to 1 byte rather than rounded up as shown in the
llvm/test/DebugInfo/X86/dwarf-aranges.ll test case.